--- pallet/src/lib.rs
@@ -1,200 +1,0 @@
-//! Aleph session manager.
-//!
-//! This pallet manages the changes in the committee responsible for establishing consensus.
-//! Currently, it's PoA where the validators are set by the root account. In the future, a new
-//! pallet for PoS elections will replace this one.
-//!
-//! For full integration with Aleph finality gadget, the `primitives::AlephSessionApi` should be implemented.
-
-#![cfg_attr(not(feature = "std"), no_std)]
-
-#[cfg(test)]
-mod mock;
-#[cfg(test)]
-mod tests;
-
-mod migrations;
-
-use sp_std::prelude::*;
-
-use frame_support::{
-    sp_runtime::BoundToRuntimeAppPublic,
-    traits::{OneSessionHandler, StorageVersion},
-    Parameter,
-};
-pub use pallet::*;
-
-/// The current storage version.
-const STORAGE_VERSION: StorageVersion = StorageVersion::new(1);
-
-#[frame_support::pallet]
-pub mod pallet {
-    use super::*;
-    use frame_support::{
-        pallet_prelude::*,
-        sp_runtime::{traits::OpaqueKeys, RuntimeAppPublic},
-    };
-    use frame_system::pallet_prelude::{ensure_root, BlockNumberFor, OriginFor};
-    use pallet_session::Pallet as Session;
-    use primitives::{
-        ApiError as AlephApiError, DEFAULT_MILLISECS_PER_BLOCK, DEFAULT_SESSION_PERIOD,
-    };
-
-    #[pallet::storage]
-    #[pallet::getter(fn validators)]
-    pub type Validators<T: Config> = StorageValue<_, Vec<T::AccountId>, OptionQuery>;
-
-    #[pallet::storage]
-    #[pallet::getter(fn session_for_validators_change)]
-    pub type SessionForValidatorsChange<T: Config> = StorageValue<_, u32, OptionQuery>;
-
-    #[pallet::config]
-    pub trait Config: frame_system::Config + pallet_session::Config {
-        type AuthorityId: Member
-            + Parameter
-            + RuntimeAppPublic
-            + Default
-            + MaybeSerializeDeserialize;
-        type Event: From<Event<Self>> + IsType<<Self as frame_system::Config>::Event>;
-    }
-
-    #[pallet::event]
-    #[pallet::generate_deposit(pub(super) fn deposit_event)]
-    pub enum Event<T: Config> {
-        ChangeValidators(Vec<T::AccountId>, u32),
-    }
-
-    #[pallet::pallet]
-    #[pallet::storage_version(STORAGE_VERSION)]
-    pub struct Pallet<T>(_);
-
-    #[pallet::hooks]
-    impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {
-        fn on_runtime_upgrade() -> frame_support::weights::Weight {
-            migrations::v0_to_v1::migrate::<T, Self>()
-        }
-    }
-
-    #[pallet::call]
-    impl<T: Config> Pallet<T> {
-        #[pallet::weight((T::BlockWeights::get().max_block, DispatchClass::Operational))]
-        pub fn change_validators(
-            origin: OriginFor<T>,
-            validators: Vec<T::AccountId>,
-            session_for_validators_change: u32,
-        ) -> DispatchResult {
-            ensure_root(origin)?;
-            Validators::<T>::put(validators.clone());
-            SessionForValidatorsChange::<T>::put(session_for_validators_change);
-            Self::deposit_event(Event::ChangeValidators(
-                validators,
-                session_for_validators_change,
-            ));
-            Ok(())
-        }
-    }
-
-    #[pallet::storage]
-    #[pallet::getter(fn authorities)]
-    pub(super) type Authorities<T: Config> = StorageValue<_, Vec<T::AuthorityId>, ValueQuery>;
-
-    #[pallet::type_value]
-    pub(super) fn DefaultForSessionPeriod() -> u32 {
-        DEFAULT_SESSION_PERIOD
-    }
-
-    #[pallet::storage]
-    #[pallet::getter(fn session_period)]
-    pub(super) type SessionPeriod<T: Config> =
-        StorageValue<_, u32, ValueQuery, DefaultForSessionPeriod>;
-
-    #[pallet::type_value]
-    pub(super) fn DefaultForMillisecsPerBlock() -> u64 {
-        DEFAULT_MILLISECS_PER_BLOCK
-    }
-
-    #[pallet::storage]
-    #[pallet::getter(fn millisecs_per_block)]
-    pub(super) type MillisecsPerBlock<T: Config> =
-        StorageValue<_, u64, ValueQuery, DefaultForMillisecsPerBlock>;
-
-    #[pallet::genesis_config]
-    pub struct GenesisConfig<T: Config> {
-        pub authorities: Vec<T::AuthorityId>,
-        pub session_period: u32,
-        pub millisecs_per_block: u64,
-    }
-
-    #[cfg(feature = "std")]
-    impl<T: Config> Default for GenesisConfig<T> {
-        fn default() -> Self {
-            Self {
-                authorities: Vec::new(),
-                session_period: DEFAULT_SESSION_PERIOD,
-                millisecs_per_block: DEFAULT_MILLISECS_PER_BLOCK,
-            }
-        }
-    }
-
-    #[pallet::genesis_build]
-    impl<T: Config> GenesisBuild<T> for GenesisConfig<T> {
-        fn build(&self) {
-            <SessionPeriod<T>>::put(&self.session_period);
-            <MillisecsPerBlock<T>>::put(&self.millisecs_per_block);
-        }
-    }
-
-    impl<T: Config> Pallet<T> {
-        pub(crate) fn initialize_authorities(authorities: &[T::AuthorityId]) {
-            if !authorities.is_empty() {
-                assert!(
-                    <Authorities<T>>::get().is_empty(),
-                    "Authorities are already initialized!"
-                );
-                <Authorities<T>>::put(authorities);
-            }
-        }
-
-        pub(crate) fn update_authorities(authorities: &[T::AuthorityId]) {
-            <Authorities<T>>::put(authorities);
-        }
-
-        pub fn next_session_authorities() -> Result<Vec<T::AuthorityId>, AlephApiError> {
-            Session::<T>::queued_keys()
-                .iter()
-                .map(|(_, key)| key.get(T::AuthorityId::ID).ok_or(AlephApiError::DecodeKey))
-                .collect::<Result<Vec<T::AuthorityId>, AlephApiError>>()
-        }
-    }
-
-    #[derive(Debug)]
-    pub enum Error {}
-
-    impl<T: Config> BoundToRuntimeAppPublic for Pallet<T> {
-        type Public = T::AuthorityId;
-    }
-
-    impl<T: Config> OneSessionHandler<T::AccountId> for Pallet<T> {
-        type Key = T::AuthorityId;
-
-        fn on_genesis_session<'a, I: 'a>(validators: I)
-        where
-            I: Iterator<Item = (&'a T::AccountId, T::AuthorityId)>,
-            T::AccountId: 'a,
-        {
-            let (_, authorities): (Vec<_>, Vec<_>) = validators.unzip();
-            Self::initialize_authorities(authorities.as_slice());
-        }
-
-        fn on_new_session<'a, I: 'a>(_changed: bool, validators: I, _queued_validators: I)
-        where
-            I: Iterator<Item = (&'a T::AccountId, T::AuthorityId)>,
-            T::AccountId: 'a,
-        {
-            let (_, authorities): (Vec<_>, Vec<_>) = validators.unzip();
-            Self::update_authorities(authorities.as_slice());
-        }
-
-        fn on_disabled(_validator_index: u32) {}
-    }
-}
